[495365] Creation of response files to manage linking and including
CMake uses normally response files which contains all the objects to
link, instead of adding them in a long command. But when cross-compiling
(apparently just from windows) it doesn't use response files. It was
added the specific commands to generate the response files, and it
worked in a tests in windows and ubuntu.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=495365
Signed-off-by: Jose Cabral <cabral@fortiss.org>