[499535] Fix warnings from compilation and clean output
The main warning is given by the negative number -9223372036854775808LL.
When parsing the compiler finds 2 tokens: the - sign and the number.
Since the number is too big, it cannot apply the sign to it, so it is
changed to: -9223372036854775807LL - 1. The output of the FB tests is
clearer now, without so many *** lines.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=499535
Signed-off-by: Jose Cabral <cabral@fortiss.org>