[500577] Integrated Lua interpreter for execution of basic function blocks
Adds a CLuaEngine class to encapsulate the Lua interpreter,
a CLuaFBTypeEntry to represent Lua function block type entries
and a CLuaFB to represent Lua function block instances.
Also adapts the CMake configuration to include the Lua interpreter.

Additionally makes some modifications to hold a Lua engine in the function block container
and some adjustments to the type library and basic function block for compatibility.

Bug: #500577
Change-Id: I5d3a51ab0e0f7eaa577a0305c5693aaac01843ed
Signed-off-by: Martin Jobst <martin.jobst@fortiss.org>