[507281] Fixes Win32SerComLayer to accept char* payload
The Win32 serial comm layer awaited a CIEC_STRING object, but due to
changes to other com layers, which changed this to pure payload to be
provided to the layer, the Win32SerComLayer failed with a segfault.
This commit provides the fix for that, as well as enhanced parameter
settings for serial interfaces, as well as the separation of parameter
parsing to a separate class.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507281
Signed-off-by: CapXilinx <melik-merkumians@acin.tuwien.ac.at>