[507431] Add bus handler for emBRICK
Add initial version of the main bus handler for emBRICK. It runs in a
seperated thread and is a singleton. It provides the basic communication
layer for the emBRICK bus protocol. The specific packages and protocol
structures are defined in the slave handler. The current implementation
approach communicates with all slaves in a constant frequency.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: I203390f906f3db3ad658454a119523986125002b
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>