[507431] Add init sequence and multiple improvements
Add the first logic of the init sequence for the emBRICK bus. The
sequential slave search is controlled by the BusHandler. The logical
packages are built and handled in the slave instances. The Packages
namespace contains all package structs for the bus communication.
Add multiple TODO statements to set guidelines for the upcoming
development.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: I48a8058448e3bc29c6dfb221edd9c3c401df3181
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>