[507431] Add data handlers and process interface
Add sync operations to the BusHandler. ProcessInterface or future config
function blocks can now check if the handler is ready.
Replace individual slave classes with slave handlers. The slave handlers
are directly linked with the slave image and are shared with process
interfaces. There exist input and output slave handlers. The input slave
handlers are continuously checked for changes. An indication event will
be fired if a change is detected. The slave handles will be created in
the config function blocks.
Add basic process interface for QX and IX function blocks.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: I2f905e2b1eef0cd0a2f08cc332ba6573e31f3c1e
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>