[507431] Sync SlaveHandles and add observer
The SlaveHandles now use a slave mutex to synchronize the manipulations
between different threads. The ProcessInterface adds itself as an
observer of an SlaveHandle and therefore listens to change events. The
slave fires the external event handler in case of a change event.

Replace the realtime clock with the monotonic clock feature. This avoids
possible errors after an unusual manipulation of the device clock.

Use a static buffer in the BusHandler to speed up transfer operations.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: I241780a0a27f4b24efc17b742be875756d63b07f
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>