[507431] Add initial version of function blocks
Add initial version of the emBRICK FORTE function blocks.
The master block handles the global configuration of the system and
initiates the configuration of the slaves. The slaves are connected via
the BusAdapter. The BusAdapter sends an init event with an initial index
of the slave. The Slave block is the parent class for all future slave
function blocks. It forwards e.g. the init event to the next slave.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: I2d6dff81e3391e77f178d1f75206b5ffc9b40021
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>