[507431] Add and integrate IOMapper
Add IOMapper singleton. The mapper works as bridge between the process
interfaces (IOObserver) and the emBRICK slave handles (IOHandle). The
IOMapper is an abstract implementation and may be used by other FORTE
modules. A device registers all its inputs and outputs as IO handles.
The runtime registers all process interfaces as IO observers. The
IOMapper maps the provided ids and may connect them. A connection is
confirmed with the onHandle and onObserver statements. In case of a
disconnection, the dropHandle and dropObserver statements are called.
Therefore the IOMapping may be reconfigured dynamically.

The process interface no longer communicates with the BusHandler. It
communicates with the IOMapper and receives an handle by the mapper. The
instance name is used as id for the IOMapper.
The master configuration block now works as controller of the
BusHandler. It starts and stops the BusHandler. The specific slave type
blocks registers all IOs at the IOMapper. The ids of the IOS are
configured by the data inputs of the slave function blocks.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: I05d9d7af7d230613df94c5d80d01a30049ce88d1
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>