[507431] Support multiple HandleTypes and improve config
Add support for words and doublewords in the IOHandle. Consequently the
ProcessInterface supports IW, QW, ID and QD function blocks.
Add SlaveHandles for analog 10bit values and 16bit values.

Add an INITO event to the configuration BusAdapter. The Master now
confirms the initialization with an INITO+/- event. A Slave can be
reconfigured during runtime. With the MAP event, the existing
connections are removed and reconnected.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: Idcb0b4d80e4106f744c4bd358e810c5af8a2815e
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>