[507431] Add error handling and improve mapping
Add error handling to all handlers. The BusHandler checks the error
status of the its subhandlers. The Master configration block checks the
status of the BusHandler.
Replace the cyclic initialization waiting loop with external events. The
Master now receives events in case the BusHandler changed state. The
Master reinitialises the BusHandler if an error occured. It sends an
INITO- after three unsuccessful attempts.
Refactor the IOHandles and SlaveHandles. Add a direction property to the
IOHandle. Add check for duplicates in the IOMapper.
Increase speed of SPIHandler after bus initialization.
Replace NULL statements with zeros.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: Id022791b3787c409d229e5760002e5def86975ac
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>