[507431] Add dynamic scheduling and sync operations
Add dynamic scheduling to the BusHandler. It replaces the cyclic
routine. Every slave has a deadline for its next update. The handles
uses pthread conditional timed waiting. Therefore it can be interrupted
by forced updates. Consequently the slaves are updated immediately after
a write request.
Allow a certain number of update errors of slaves. If the maximum of
errors is reached, the BusHandlers reinitializes the bus.
Add sync mutexes to process interfaces and configuration function
blocks. The IOMapper operations are not threadsafe. Therefore the events
must be protected by sync objects.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: I704593c48143a4f0b1a7e64877cba1d80f724595
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>