[507431] Refactor and multiple improvements
Fix segfault in forceUpdate of BusHandler.
Refactor ProcessInterface. Replace specific methods with two read and
write methods.
Init and reset SlaveHandles correctly. The current state is set after
mapping. The handle resets if the mapping is disconnected.
Handle the status code of the emBrick modules. The status is accessible
via the status data output of the configuration function blocks.
Add support for the 2-301 (6 Relays) emBrick module.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: I2e32acfda1a7176ccec7f574f92d331863685939
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>