[511632] Closing connections on Killing CommFbs
So far ComFBs closed their connection only on INIT- and on the 
deletion of the FB. This is rather late and can lead to 
unwanted race conditions. With this patch the connection is
also closed when the CommFB is killed. This first of all allows
a better reseting reinitialization and furthermore should 
reduce the potential risk of the race condition identified in 
the referenced bug.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=511632
Signed-off: alois.zoitl@gmx.at