[507431] Refactor config FBs and IO Mapper
Adapt configuration function blocks to the thesis defintion. Add
configuration parameters and the correct events. Fix error handling of
the EBMaster. Add default parameters and helper methods.
Add reference of the Slave instance to the SlaveHandel class. Add
forceUpdate method to the Slave class.
Refactor Mapper classes. Add Direction enumeration to Mapper and extend
the Observer with type and direction.
Refactor the ProcessInterface. Use the direction property of the
Observer class.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=507431
Change-Id: I9abe5dec26cae46dca7a7d53e2e14a9426868cc3
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>