[514372] Com layers creation code redistribution
Moving the recursive creation of all the layers to CCommFB avoids
unneccesary complexity and function call in CComLayer. A plain while
loop has been used for this purpouse instead of a recursive function. If
an error happens during this process, the previously created layers are
deleted. ID parsing functions have been moved to CComLayersManager in
order to clean CCommFB.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=514372
Change-Id: I7048879d9ade7cd2a7510ed8b601771267358363
Signed-off-by: Adrián Orive <aorive@ikerlan.es>