[507431] Add controller for IOs with poll operations
IO device controller for devices which require an implementation of IOs
using poll operations. Offers a poll method which performs an IO update
in a configured PollInterval. Allows to force a polling routine with the
forcePoll method (e.g. can be used to set an output immediately).

Add libary realtime (rt) to posix architecture. Required for timing
operations.

Change-Id: If47763e48616b8d5fa8094a2410a9a9483b73c06
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>