[507431] Move IO lib and add Handle descriptors
Moves the io implementation from the fblib to the core folder. The
embrick implementation is adapted accordingly. Additionally the
BitSlaveHandle is introduced as global lib as it is used in many IO
implementations.

Handle descriptors are introduced which are used to exchange information
about handles between the device controller and the corresponding
configuration fb. The IO implementation should extend the Handle struct
with properties, which uniquely identify the handle.

Change-Id: I3f7fb97dde072dbf347dbda7e72f1f2677cc3cfe
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>