[507431] Add support for splitted io fb networks
Add support for splitted configuration fb networks. If a single
controller has to many ios or ios with invididual configuration
parameters, the available ios can be split in multiple parts. A part is
a separate configuration fb which is connected with an individual
adapter to the main SplitController.

Add error detection to the PollController.

Change-Id: I3169f5c74781a11285dd1e2b0354d15e32dc5fc6
Signed-off-by: Johannes Meßmer <johannes.messmer@tum.de>