Enhance the bootfile loading logic
This change cleans up the bootfile loading logic a bit.
It also allows to the set the envvar FORTE_BOOT_FILE_FAIL_MISSING
to any value in order to fail forte when the bootfile is missing.
The change also allows setting FORTE_BOOT_FILE to an empty value
which disables the bootfile loading.

Signed-off-by: Jens Reimann <jreimann@redhat.com>