[525397] Get rid of static device execution
The device execution is now a reference in every external event
handler. Changes are done to instantiate all external handlers
for every device execution, and FB accessing the external event
handlers are provided with helper function to do it, by finding
it device, and from that the external handler. Changes are missing
in rcX arch. Linux, Windows and VxWorks arch were tested, mqtt,
opcua were also tested. Modbus is compiling but wasn't tested

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=525397
Change-Id: I7d75c290c1edd8230b15b285f0cf40d6726d1302
Signed-off-by: Jose Cabral <cabral@fortiss.org>