[529959] Changes POSIX to monotonic clock source
This commit modifies the posix timer handler to use the monotonic clock
source. Several changes have to be made, as clock_gettime uses the
timespec struct instead of timeval. As there are not library
implementation of subtraction, addition and comparison for timepecs I
added them plus some tests, checking the behavior of addition and
subtraction.
Corrects header information

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=529959

Signed-off-by: CapXilinx <melik-merkumians@acin.tuwien.ac.at>