[537006] Adds capability to use mixed bit type AND in ST
This fix adds the capabiltiy to use mixed bit type AND in ST and C++
code, returning the explicitly the widest used type. The return of
CIEC_ANY_BIT would result in other compilation errors, due to the fact
that it cannot be directly converted to a bit type.

Adds compile guard for incompatible types

Uses C++11 template features

Change-Id: Id1a90f6bc73061062adc8a0f4635894948a92d7d
Signed-off-by: Martin Melik Merkumians <melik-merkumians@acin.tuwien.ac.at>