[537601] First version of a base class for generic fbs
introduced base class for generic fbs to overcome the need for the
manages data flag in CFunctionBlock. Furthermore some common
functionality for handling the typename and configuring the fb has been
moved to this base class.

As the commit is already rather big I didn't cleanup all the involved
classes.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=537601
Change-Id: I51a36e5bad9a82741860846eb32984f2c024c763
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>