[538870] Fixes crash by divide by zero or undefined input
The fix supressed the execution of the F_DIV FB in the case of
non-connected inputs, and prevents a divison by zero expection. In the
case of a divide by zero the result will be set to zero, until a better
solution can be found, but prevents that FORTE crashes in these cases.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=538870
Change-Id: I2e550e849ac1e88ca314d887930cd0c105d3ff00
Signed-off-by: Martin Melik-Merkumians <melik-merkumians@acin.tuwien.ac.at>