[538391] Remove expected code from PARAMS and allow response in POST/PUT
Types of SD and RD are checked. PUT/POST can return body code, so in all
cases, the client needs 2 RD, one for the response code, and the the
second for the body. If some response code must be omited, this function
should be done in the application. Also, tests for the parser is added
and a big cleanup was done

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=538391
Change-Id: I0050c8472696095e9bdc42856602867477887963
Signed-off-by: Jose Cabral <cabral@fortiss.org>