Allow to easily add new types and remove typeConverter
Structures can easily be added in cmake. The transformation
from/to opcua/61499 is done automatically from the information
of the structure. The first base for arrays is also present,
but not ready yet. Also, the typeConverter is not public anymore
and common functios are provided and the transformation happends
according to the passed CIEC_ANY variable

Change-Id: Ib1c7ee24ada5a6099917d502f2acea9b433e65e6
Signed-off-by: Jose Cabral <cabral@fortiss.org>