[549852] Corrected adapter event id transformation
4diac FORTE currently indicates adapter events by using the adapter Id
(index in the adapter list +1) in the higher 8 bits and the event id in
the lower 8 bits. The current Lua implementation uses an Adapter Flag,
the adapter id in the higher 16 bits and the event id. The
transformation from the first to the second was missing the << 16 after
the 1 was substracted from the adapter ID.

This is only a first quick fix. I think in the mid turn run we should
consider to ommit these transformations at all. This would then only
affect the code generator and simplify debugging and common 4diac FORTE
behavior.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=549852
Change-Id: Ibde38cbfc45271435b18a5798f8ec81c74b64adc
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>