[569084] change POSIX semaphores to not busy-wait
The current implementation of semaphores on the POSIX platform causes
100% CPU load when timedWait() is called.  It is based on POSIX
semaphores and cannot be fixed due to API shortcomings.

This patch replaces the old implementation with one based on POSIX
condition variables, which allow bug-free waiting with timeout.  The
semaphore type is also changed to be a binary semaphore in order to be
in line with other platforms.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=569084
Signed-off-by: Jörg Walter <joerg.walter@offis.de>
Change-Id: I215ccbe4130096bb759db309258bd2589bd66b81