[569144] Fix FORTE_LINKED_STRINGDICT definition
A misplaced definition of build option FORTE_LINKED_STRINGDICT leads to
a superfluous recompile of all source files on the second build in a ny
given build directory. Its value is referenced before it is defined.

This patch moves the option declaration in CMakeLists.txt up so that it
can take effect before FORTE_LINKED_STRINGDICT is referenced first.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=569144
Signed-off-by: Jörg Walter <joerg.walter@offis.de>
Change-Id: Ie0b964c9e143ac96d5aa2608f6ca61a8c4b7e7fc