Bug 578876 Fixes boolean operations for partial access
Partial acccess creates specialized objects to handle the partial
access. The issue is, that the template classes require the base types
to function correct, and unfortunately the PARTIAL classes cannot be
added easily to the cast semantics, as they are nested classes of
CIEC_ANY_BIT. This commit introduces a new template metaprogram,
checking on the supertype and returning the approriate CIEC class. This
can then be used in the function template to replace the calling types
of the other templates.

Change-Id: Ie43a13d5d5577369225e6e4405fd3393d63496de
Signed-off-by: Martin Melik Merkumians <melik-merkumians@acin.tuwien.ac.at>