Bug 579171 Removes builtin type operator=
Removes the builtin operator= to give the compiler no chance to somehow cast his way arounf forbidden casts.
Also introduces a separate type for representing the state of an BFB, which needs to still be monitorable.
Also changes the escape symbol for watching the ECC state from $ECC to !ECC as $ is the STRING escape symbol.

Signed-off-by: Martin Melik-Merkumians <melik-merkumians@acin.tuwien.ac.at>
Change-Id: Ia12423ca30db4f169b29a53dd5637533408d76dd