Bug 579743 - Fixes CIEC_TIME toString
Negative fractions of ms get serialized wrong.
E.g. T#-0.1ms gets serialzed as T#0.-1ms, as the ms part is rounded to 0,
which is non-negativ, but the us part is -100us, which is a negative number
and is serialized as such after the decimal dot.

This patch first checks if the time value is negative, then prepared the T#
part of the literal, and then handles the remainder as positive number.

Signed-off-by: Martin Melik-Merkumians <melik-merkumians@acin.tuwien.ac.at>
Change-Id: I3c6eb66dfcdfaedc8ff1e471596fcf1dadcdafb0