Bug 580588 - Remove macro magic from test fixture
In classes derived from CFBTestFixtureBase, the functions setInputData()
and setOutputData() have to be called with pointers and lengths as parameters.
To simplify the usage, macros are used to set the values.

This is now simplified by using the C++11 feature std::initializer_list<>.
Since it is a lightweight wrapper for an array, the function signatures
are simplified and the need for ugly macros is removed.

Change-Id: Icc52478dd1a49a9cc07ad83bd8ebf3c876ecf920