Bug 580554 - Fixed unescaping of write command attributes
During parsing of the write command the value to write had to be
unescaped. This was done on the orginal string. However this changed the
original provided string such that parsing of the force attribute didn't
work anymore.

This change now first assigns the parsed value to the additionalParams
value and performs the unescaping savely there.

Change-Id: Idd49f957deaa77f4c9dbb44da11ca7417da9ed10