Bug 581561 - Necessary changes to enable partial as outs
PARITALs are subclasses from a specific ANY_BIT class.
When using the returned PARTIAL now in a function/method call the
assumed type is then one of the concrete ANY_BIT classes
(BOOL, BYTE, etc.). The compiler statically binds the
operator= of one of these classes for an assignment in the calling
function/method.

When now a PARTIAL is given as a parameter, the compiler has statically
bound statically operator= of the concrete non-partial class and by it
the setValueSimple is used.

The internal partial value is then updated, but its not written-through
to the original data holding variable, which spawned the partial.

Therefore the operator= needs to use the virtual setValue method, so
the correct steps are executed in case of a PARTIAL.

Signed-off-by: Martin Melik-Merkumians <melik-merkumians@acin.tuwien.ac.at>
Change-Id: I746b5e5290529d6e660186a78cbb62ad8e5e9f6f