Bug 582189 - Fixes partial string access
Partial string access now can write to unwritten string symbols,
e.g. in string "test" accessing the 6th element, which is two symbols after the final 't'
returns when read a \0, and when written the 6th element now holds the assigned value (lets assume '1').
Symbols between the accessed element and the last one in the string are filled with \0,
so the resulting string would be "test\01".

Also some bugs were fixed, which resulted in strange behavior if used as an inout or out param.

Signed-off-by: Martin Melik-Merkumians <melik-merkumians@acin.tuwien.ac.at>
Change-Id: I066ba6f88e4642d5a1672703813e438d5e99e4e4