Bug 582189 - Fix resize when reading past the end of a string
A string should not be resized when reading past the end. Checks if the
value is still '\0' before extending the string. Note that this does not
cover the case when explicitly writing a '\0' character to the string,
in which case the string will not be extended. However, reading from the
string afterwards will correctly return a '\0', despite the length
remaining untouched.

Change-Id: I445d13ef6839685a5b7f372e290b95e765cfa0a7
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>