Fix query FB instances and connections commands
A resource was previously added to the FB list of a device. With the
recent changes to unify both FB and sub-contaier lists in the FB
container, resources were no longer considered FB instances when
querying FB instances or connections. This also meant the resource
itself was no longer returned when querying the device.

This commit introduces a new virtual method isDynamicContainer() to
distinguish between FBs and/or dynamic containers, respectively. The
former are now all subclasses of CFunctionBlock and the latter support
adding or removing children at runtime.