Moved I/O FBs to Modular I/O Infrastructure
The current I/O FB implemenation is based on the original generic I/O
interface. This is not best suited for the modular I/O infrastructure.
As more and more I/O implementations move to the modular I/O it makes
sense to have a dedicated optimized I/O FB implementation for the
modular I/O infrastructure. This first commit moves the FBs there and
correctly sets up the build configuration.