[cleanup] Removed CTOR Defines
The CTOR defines was original added to hide CTOR setups to make 4diac
FORTE easier upgradeable. With the latest changes they are not realy
used anymore and do not make sense. Therefore they are removed and the
last FBs that still had them where ported to the latest setup. However,
please note that these FBs have not tested in quite some while and my
still not work.