Fix line feed and newline escape sequences in strings and chars
The '$L' or Line Feed escape sequence in (W)STRING was assigned
the hexadecimal value 0x10, which turns out to be Data Link Escape.
Both (W)STRING and (W)CHAR also escaped the character value '\n'
back to $N or Newline, which is an implementation-independent alias
for the end of a line.

This changes '$L' or Line Feed to use the value '\n' (0x0a) and also
translates '\n' back to '$L'. The escape '$N' or Newline is mapped to
the implementation-defined value '\n' on all platforms.