Add script to combine static forte libraries
Add script to combine static forte libraries for easier consumption by
non-CMake build systems. The script combines all libforte*.a files in
the current working directory to a libforte-all.a, except any existing
libforte-all.a itself. It is intended to be run inside the lib directory
of the install prefix after installing forte and all external modules.