Replaces F_MUX_2 with a generic F_MUX FB
Replaces the F_MUX_2 FB with a generic F_MUX FB that can handle a variable number of inputs.
This allows for more flexibility and reusability of the function block.

.

.

Removes unused setInitialValues function

Removes the `setInitialValues` function from the `GEN_F_MUX` class, as it is no longer needed.

Updates logging to use debug level instead of info.